+++ ThirdPartyAdapters/mytarget/mytarget/src/main/java/com/google/ads/mediation/mytarget/MyTargetAdapter.java
@@ -6,7 +6,7 @@ import android.util.Log;
 import android.view.View;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import com.google.android.gms.ads.AdRequest;
+import com.google.android.gms.ads.AdError;
 import com.google.android.gms.ads.AdSize;
 import com.google.android.gms.ads.mediation.MediationAdRequest;
 import com.google.android.gms.ads.mediation.MediationBannerAdapter;
@@ -17,7 +17,6 @@ import com.my.target.ads.InterstitialAd;
 import com.my.target.ads.MyTargetView;
 import com.my.target.ads.MyTargetView.MyTargetViewListener;
 import com.my.target.common.CustomParams;
-
 import java.util.Date;
 import java.util.GregorianCalendar;
 
@@ -43,22 +42,24 @@ public class MyTargetAdapter extends MyTargetMediationAdapter
       AdSize adSize,
       MediationAdRequest mediationAdRequest,
       Bundle mediationExtras) {
+
     int slotId = MyTargetTools.checkAndGetSlotId(context, serverParameters);
-    Log.d(TAG, "Requesting myTarget banner mediation, slotId: " + slotId);
+    Log.d(TAG, "Requesting myTarget banner mediation with Slot ID: " + slotId);
+
     if (slotId < 0) {
-      if (mediationBannerListener != null) {
-        mediationBannerListener.onAdFailedToLoad(
-            MyTargetAdapter.this, AdRequest.ERROR_CODE_INVALID_REQUEST);
-      }
+      AdError error = new AdError(ERROR_INVALID_SERVER_PARAMETERS, "Missing or invalid Slot ID.",
+          ERROR_DOMAIN);
+      Log.e(TAG, error.getMessage());
+      mediationBannerListener.onAdFailedToLoad(MyTargetAdapter.this, error);
       return;
     }
 
-    if (adSize == null) {
-      Log.w(TAG, "Failed to request ad, AdSize is null.");
-      if (mediationBannerListener != null) {
-        mediationBannerListener.onAdFailedToLoad(
-            MyTargetAdapter.this, AdRequest.ERROR_CODE_INVALID_REQUEST);
-      }
+    MyTargetView.AdSize myTargetSize = MyTargetTools.getSupportedAdSize(adSize, context);
+    if (myTargetSize == null) {
+      String errorMessage = String.format("Unsupported ad size: %s.", adSize.toString());
+      AdError error = new AdError(ERROR_BANNER_SIZE_MISMATCH, errorMessage, ERROR_DOMAIN);
+      Log.e(TAG, error.getMessage());
+      mediationBannerListener.onAdFailedToLoad(MyTargetAdapter.this, error);
       return;
     }
 
@@ -66,45 +67,12 @@ public class MyTargetAdapter extends MyTargetMediationAdapter
     if (mediationBannerListener != null) {
       bannerListener = new MyTargetBannerListener(mediationBannerListener);
     }
-    int width = adSize.getWidth();
-    int height = adSize.getHeight();
-    int widthInPixels = adSize.getWidthInPixels(context);
-    int heightInPixels = adSize.getHeightInPixels(context);
-    if (width == 300 && height == 250) {
-      Log.d(TAG, "Loading myTarget banner, size: 300x250");
-      loadBanner(bannerListener,
-          mediationAdRequest,
-          slotId,
-          MyTargetView.AdSize.ADSIZE_300x250,
-          context);
-    } else if (adSize.getWidth() == 728 && adSize.getHeight() == 90) {
-      Log.d(TAG, "Loading myTarget banner, size: 728x90");
-      loadBanner(bannerListener,
-          mediationAdRequest,
-          slotId,
-          MyTargetView.AdSize.ADSIZE_728x90,
-          context);
-    } else if (adSize.getWidth() == 320 && adSize.getHeight() == 50) {
-      Log.d(TAG, "Loading myTarget banner, size: 320x50");
-      loadBanner(bannerListener,
-          mediationAdRequest,
-          slotId,
-          MyTargetView.AdSize.ADSIZE_320x50,
-          context);
-    } else if (widthInPixels > 0 && heightInPixels > 0 && heightInPixels < 0.75f * widthInPixels) {
-      Log.d(TAG, "Loading myTarget banner, size: adaptive");
-      loadBanner(bannerListener,
-                 mediationAdRequest,
-                 slotId,
-                 MyTargetView.AdSize.getAdSizeForCurrentOrientation(MyTargetTools.toDips(widthInPixels, context), context),
-                 context);
-    } else {
-      Log.w(TAG, "AdSize " + adSize.toString() + " is not currently supported");
-      if (mediationBannerListener != null) {
-        mediationBannerListener
-            .onAdFailedToLoad(MyTargetAdapter.this, AdRequest.ERROR_CODE_NO_FILL);
-      }
-    }
+
+    String logMessage = String
+        .format("Loading myTarget banner with size: %dx%d.", myTargetSize.getWidth(),
+            myTargetSize.getHeight());
+    Log.d(TAG, logMessage);
+    loadBanner(bannerListener, mediationAdRequest, slotId, myTargetSize, context);
   }
 
   @Override
@@ -119,20 +87,18 @@ public class MyTargetAdapter extends MyTargetMediationAdapter
       MediationAdRequest mediationAdRequest,
       Bundle mediationExtras) {
     int slotId = MyTargetTools.checkAndGetSlotId(context, serverParameters);
-    Log.d(TAG, "Requesting myTarget interstitial mediation, slotId: " + slotId);
+    Log.d(TAG, "Requesting myTarget interstitial mediation with Slot ID: " + slotId);
 
     if (slotId < 0) {
-      if (mediationInterstitialListener != null) {
-        mediationInterstitialListener.onAdFailedToLoad(
-            MyTargetAdapter.this, AdRequest.ERROR_CODE_INVALID_REQUEST);
-      }
+      AdError error = new AdError(ERROR_INVALID_SERVER_PARAMETERS, "Missing or invalid Slot ID.",
+          ERROR_DOMAIN);
+      Log.e(TAG, error.getMessage());
+      mediationInterstitialListener.onAdFailedToLoad(MyTargetAdapter.this, error);
       return;
     }
 
-    MyTargetInterstitialListener bannerListener = null;
-    if (mediationInterstitialListener != null) {
-      bannerListener = new MyTargetInterstitialListener(mediationInterstitialListener);
-    }
+    MyTargetInterstitialListener interstitialListener = new MyTargetInterstitialListener(
+        mediationInterstitialListener);
 
     if (mInterstitial != null) {
       mInterstitial.destroy();
@@ -160,7 +126,7 @@ public class MyTargetAdapter extends MyTargetMediationAdapter
         }
       }
     }
-    mInterstitial.setListener(bannerListener);
+    mInterstitial.setListener(interstitialListener);
     mInterstitial.load();
   }
 
@@ -193,10 +159,10 @@ public class MyTargetAdapter extends MyTargetMediationAdapter
    * Starts loading banner.
    *
    * @param myTargetBannerListener listener for ad callbacks
-   * @param mediationAdRequest Google mediation request
-   * @param slotId myTarget slot ID
-   * @param adSize myTarget banner size.
-   * @param context app context
+   * @param mediationAdRequest     Google mediation request
+   * @param slotId                 myTarget slot ID
+   * @param adSize                 myTarget banner size.
+   * @param context                app context
    */
   private void loadBanner(@Nullable MyTargetBannerListener myTargetBannerListener,
       @Nullable MediationAdRequest mediationAdRequest,
@@ -254,24 +220,25 @@ public class MyTargetAdapter extends MyTargetMediationAdapter
 
     @Override
     public void onLoad(@NonNull final MyTargetView view) {
-      Log.d(TAG, "Banner mediation Ad loaded");
+      Log.d(TAG, "Banner mediation Ad loaded.");
       listener.onAdLoaded(MyTargetAdapter.this);
     }
 
     @Override
     public void onNoAd(@NonNull final String reason, @NonNull final MyTargetView view) {
-      Log.i(TAG, "Banner mediation Ad failed to load: " + reason);
-      listener.onAdFailedToLoad(MyTargetAdapter.this, AdRequest.ERROR_CODE_NO_FILL);
+      AdError error = new AdError(ERROR_MY_TARGET_SDK, reason, MY_TARGET_SDK_ERROR_DOMAIN);
+      Log.e(TAG, error.getMessage());
+      listener.onAdFailedToLoad(MyTargetAdapter.this, error);
     }
 
     @Override
     public void onShow(@NonNull MyTargetView view) {
-      Log.d(TAG, "Banner mediation Ad show");
+      Log.d(TAG, "Banner mediation Ad show.");
     }
 
     @Override
     public void onClick(@NonNull final MyTargetView view) {
-      Log.d(TAG, "Banner mediation Ad clicked");
+      Log.d(TAG, "Banner mediation Ad clicked.");
       listener.onAdClicked(MyTargetAdapter.this);
       listener.onAdOpened(MyTargetAdapter.this);
       // click redirects user to Google Play, or web browser, so we can notify
@@ -294,19 +261,20 @@ public class MyTargetAdapter extends MyTargetMediationAdapter
 
     @Override
     public void onLoad(@NonNull final InterstitialAd ad) {
-      Log.d(TAG, "Interstitial mediation Ad loaded");
+      Log.d(TAG, "Interstitial mediation Ad loaded.");
       listener.onAdLoaded(MyTargetAdapter.this);
     }
 
     @Override
     public void onNoAd(@NonNull final String reason, @NonNull final InterstitialAd ad) {
-      Log.i(TAG, "Interstitial mediation Ad failed to load: " + reason);
-      listener.onAdFailedToLoad(MyTargetAdapter.this, AdRequest.ERROR_CODE_NO_FILL);
+      AdError error = new AdError(ERROR_MY_TARGET_SDK, reason, MY_TARGET_SDK_ERROR_DOMAIN);
+      Log.e(TAG, error.getMessage());
+      listener.onAdFailedToLoad(MyTargetAdapter.this, error);
     }
 
     @Override
     public void onClick(@NonNull final InterstitialAd ad) {
-      Log.d(TAG, "Interstitial mediation Ad clicked");
+      Log.d(TAG, "Interstitial mediation Ad clicked.");
       listener.onAdClicked(MyTargetAdapter.this);
       // click redirects user to Google Play, or web browser, so we can notify
       // about left application.
@@ -315,7 +283,7 @@ public class MyTargetAdapter extends MyTargetMediationAdapter
 
     @Override
     public void onDismiss(@NonNull final InterstitialAd ad) {
-      Log.d(TAG, "Interstitial mediation Ad dismissed");
+      Log.d(TAG, "Interstitial mediation Ad dismissed.");
       listener.onAdClosed(MyTargetAdapter.this);
     }
 
@@ -325,7 +293,7 @@ public class MyTargetAdapter extends MyTargetMediationAdapter
 
     @Override
     public void onDisplay(@NonNull final InterstitialAd ad) {
-      Log.d(TAG, "Interstitial mediation Ad displayed");
+      Log.d(TAG, "Interstitial mediation Ad displayed.");
       listener.onAdOpened(MyTargetAdapter.this);
     }
   }