+++ ThirdPartyAdapters/verizonmedia/verizonmedia/src/main/java/com/google/ads/mediation/verizon/AdapterUnifiedNativeAdMapper.java
@@ -100,7 +100,7 @@ class AdapterUnifiedNativeAdMapper extends UnifiedNativeAdMapper {
 
   @Override
   public void recordImpression() {
-    verizonAd.fireImpression();
+    verizonAd.fireImpression(context);
   }
 
   @Override