+++ ThirdPartyAdapters/ironsource/CHANGELOG.md
@@ -1,5 +1,43 @@
 ## IronSource Android Mediation Adapter Changelog
 
+#### Version 7.1.0.2.0
+- Verified compatibility with ironSource SDK version 7.1.0.2.
+
+Built and tested with:
+- Google Mobile Ads SDK version 19.7.0.
+- IronSource SDK version 7.1.0.2.
+
+#### Version 7.1.0.1.0
+- Verified compatibility with ironSource SDK version 7.1.0.1.
+- Updated the minimum required Google Mobile Ads SDK version to 19.7.0.
+
+Built and tested with:
+- Google Mobile Ads SDK version 19.7.0.
+- IronSource SDK version 7.1.0.1.
+
+#### Version 7.0.4.1.0
+- Verified compatibility with ironSource SDK version 7.0.4.1.
+- Updated the minimum required Google Mobile Ads SDK version to 19.6.0.
+
+Built and tested with:
+- Google Mobile Ads SDK version 19.6.0.
+- IronSource SDK version 7.0.4.1.
+
+#### Version 7.0.4.0
+- Verified compatibility with ironSource SDK version 7.0.4.
+
+Built and tested with:
+- Google Mobile Ads SDK version 19.5.0.
+- IronSource SDK version 7.0.4.
+
+#### Version 7.0.3.1.0
+- Verified compatibility with ironSource SDK version 7.0.3.1.
+- Updated the minimum required Google Mobile Ads SDK version to 19.5.0.
+
+Built and tested with:
+- Google Mobile Ads SDK version 19.5.0.
+- IronSource SDK version 7.0.3.1.
+
 #### Version 7.0.2.0
 - Verified compatibility with ironSource SDK version 7.0.2.
 - Updated the minimum required Google Mobile Ads SDK version to 19.4.0.